+++ shared/src/commonMain/kotlin/site/remlit/snowdrop/component/Emoji.kt
@@ -2,7 +2,9 @@ package site.remlit.snowdrop.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
@@ -13,7 +15,8 @@ import io.kamel.image.KamelImage
 import io.kamel.image.asyncPainterResource
 import site.remlit.snowdrop.model.Emoji
 
-const val emojiSize = 20
+const val emojiSize = 24
+const val mediumEmojiSize = 36
 const val bigEmojiSize = 40
 
 /**
@@ -27,6 +30,7 @@ const val bigEmojiSize = 40
 @Composable
 fun Emoji(
 	emoji: Emoji,
+	fill: Boolean = false,
 	big: Boolean = false
 ) {
 	val size = if (big) bigEmojiSize.dp else emojiSize.dp
@@ -35,8 +39,10 @@ fun Emoji(
 	fun fallback() {
 		Box(
 			modifier = Modifier.background(MaterialTheme.colorScheme.surfaceContainerHigh)
-				.height(size)
-				.width(size)
+				.let {
+					if (fill) it.fillMaxSize()
+					else it.size(size)
+				}
 		)
 	}
 
@@ -45,7 +51,9 @@ fun Emoji(
 		contentDescription = emoji.shortcode,
 		contentScale = ContentScale.Fit,
 		onLoading = { fallback() },
-		modifier = Modifier.height(size)
-			.width(size),
+		modifier = Modifier.let {
+			if (fill) it.fillMaxSize()
+			else it.size(size)
+		},
 	)
 }